Fixed CMake toolchain for M1 crossbuilding and moved/completed test
